+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.2
 
+- Enable default designation through service tagging
 - Add a Docker Compose stack for testing
 - Add a middleware for automatically inserting the `X-Country-Code` header